requirements.txt
alembic==1.7.5
aniso8601==9.0.1
app==0.0.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-WTF==0.15.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.6
MarkupSafe==2.0.1
passlib==1.7.4
Pillow==8.4.0
psycopg2==2.9.3
pycparser==2.21
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.26
typing-extensions==3.10.0.2
visitor==0.1.3
Werkzeug==2.0.2
WTForms==2.3.3
zipp==3.6.0
